nominal changes

+++ src/utils/StorageManager/setup.ts
@@ -9,6 +9,9 @@ import { AppStorageManager } from './repositories/appStorageManager';
 import { AssetStorageRepository } from './repositories/assetStorageRepository';
 import { AssetStorageManager } from './repositories/assetStorageManager';
 import { AssetStorage } from './entities/assetStorage';
+import Debug from 'debug';
+
+const debug = Debug('tonomy-id:utils:StorageManage:setup');
 
 export const dataSource = new DataSource({
     database: 'storage',
@@ -55,25 +58,38 @@ async function checkTableExists(dataSource, tableName) {
     return result.length > 0;
 }
 
-async function resetAssetTable() {
+async function checkColumnExists(dataSource, tableName, columnName) {
+    const queryRunner = dataSource.createQueryRunner();
+    const result = await queryRunner.query(`PRAGMA table_info(${tableName})`);
+
+    await queryRunner.release();
+    return result.some((column) => column.name === columnName);
+}
+
+async function resetAssetTableIfColumnMissing() {
     const queryRunner = dataSource.createQueryRunner();
 
     await queryRunner.connect();
 
     try {
-        // Drop the AssetRecordStorage table
-        await queryRunner.query('DROP TABLE IF EXISTS "AssetStorage"');
-        console.log('AssetStorage table dropped successfully.');
+        const columnExists = await checkColumnExists(dataSource, 'AssetStorage', 'assetName');
 
-        // Synchronize the schema
-        if (!dataSource.isInitialized) {
-            await dataSource.initialize();
+        if (!columnExists) {
+            await queryRunner.query('DROP TABLE IF EXISTS "AssetStorage"');
+            debug('AssetStorage table dropped successfully.');
+
+            if (!dataSource.isInitialized) {
+                await dataSource.initialize();
+            }
+
+            await dataSource.synchronize();
+            debug('Schema synchronized successfully.');
+        } else {
+            // Drop the AssetStorage table
+            console.log('AssetStorage table already has the assetName column.');
         }
-
-        await dataSource.synchronize();
-        console.log('Schema synchronized successfully.');
     } catch (error) {
-        console.error('Error resetting AssetStorage table:', error);
+        debug('Error checking or resetting AssetStorage table:', error);
     } finally {
         await queryRunner.release();
     }
@@ -99,6 +115,6 @@ export async function connect() {
 }
 
 // Call resetAssetTable to drop the AssetRecordStorage table and synchronize schema
-resetAssetTable().catch((error) => {
+resetAssetTableIfColumnMissing().catch((error) => {
     console.error('Error in resetAssetTable:', error);
 });
